refactor: use KakarotTestEnvironmentContext instead of KakarotTestEnvironment
