+++ crates/core/Cargo.toml
@@ -42,6 +42,7 @@ ethers = { workspace = true }
 ethers-solc = { workspace = true }
 foundry-config = { git = "https://github.com/foundry-rs/foundry", branch = "master" }
 walkdir = "2.3.3"
+rstest = { workspace = true }
 
 [dev-dependencies]
 dojo-test-utils = { workspace = true }
@@ -51,4 +52,3 @@ tracing-subscriber = "0.3.17"
 tracing = "0.1.37"
 ctor = "0.2.4"
 cargo-husky = { workspace = true }
-rstest = { workspace = true }
