+++ crates/core/src/test_utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod constants;
 pub mod deploy_helpers;
+pub mod fixtures;
